all migrations added and tested

+++ migrations/20210713094615-users.js
@@ -5,28 +5,48 @@ var type;
 var seed;
 
 /**
-  * We receive the dbmigrate dependency from dbmigrate initially.
-  * This enables us to not have to rely on NODE_PATH.
-  */
-exports.setup = function(options, seedLink) {
+ * We receive the dbmigrate dependency from dbmigrate initially.
+ * This enables us to not have to rely on NODE_PATH.
+ */
+exports.setup = function (options, seedLink) {
   dbm = options.dbmigrate;
   type = dbm.dataType;
   seed = seedLink;
 };
 
-exports.up = function(db) {
+exports.up = function (db) {
   return db.createTable("users", {
-      id: {type:"smallint", primaryKey: true, autoIncrement: true},
-      username: {type: "string", length: 45, notNull:false},
-      email: {type:"string", length:100},
-      password: {type:"string", length: 100},
-      phone:{type:"smallint", unsigned: true},
-      address: {type:"string", length: 255},
-      dob: "date"
+    id: {
+      type: "int",
+      primaryKey: true,
+      autoIncrement: true
+    },
+    username: {
+      type: "string",
+      length: 45,
+      notNull: false
+    },
+    email: {
+      type: "string",
+      length: 100
+    },
+    password: {
+      type: "string",
+      length: 100
+    },
+    phone: {
+      type: "int",
+      unsigned: true
+    },
+    address: {
+      type: "string",
+      length: 255
+    },
+    dob: "date"
   });
 };
 
-exports.down = function(db) {
+exports.down = function (db) {
   return db.dropTable("users");
 };
